LRAC-14780 Simplify
